REBASE OLLAMA CONFIG OUT
break out @runt/ai

I have no internet so I'm cleaning up old fast prototype gunk. To get it
working with no internets, I'm using Ollama slightly hard coded.
