+++ deno.lock
@@ -1009,6 +1009,17 @@
   },
   "workspace": {
     "members": {
+      "packages/ai": {
+        "dependencies": [
+          "jsr:@openai/openai@^4.98.0",
+          "jsr:@runt/lib@~0.4.2",
+          "jsr:@runt/schema@~0.4.2",
+          "jsr:@std/async@1",
+          "npm:@livestore/livestore@~0.3.1",
+          "npm:pyodide@~0.27.7",
+          "npm:strip-ansi@^7.1.0"
+        ]
+      },
       "packages/lib": {
         "dependencies": [
           "jsr:@runt/schema@~0.4.2",
@@ -1022,6 +1033,7 @@
       "packages/pyodide-runtime-agent": {
         "dependencies": [
           "jsr:@openai/openai@^4.98.0",
+          "jsr:@runt/ai@~0.4.2",
           "jsr:@runt/lib@~0.4.2",
           "jsr:@runt/schema@~0.4.2",
           "jsr:@std/async@1",
