+++ packages/pyodide-runtime-agent/deno.json
@@ -15,6 +15,7 @@
     "pyrunt": "./src/mod.ts"
   },
   "imports": {
+    "@runt/ai": "jsr:@runt/ai@^0.4.2",
     "@runt/lib": "jsr:@runt/lib@^0.4.2",
     "@runt/schema": "jsr:@runt/schema@^0.4.2",
     "npm:pyodide": "npm:pyodide@^0.27.7",
